created custom hooks (sorting, api calling)
